clip-scraper/requirements.txt
fastapi==0.110.1
uvicorn[standard]==0.25.0
httpx==0.26.0
playwright>=1.40.0
python-dotenv==1.0.1
pydantic==2.5.3
